refactor: extract clock test assertions into shared helpers in common.rs
