+++ crates/core-node-internal/tests/common.rs
@@ -7,15 +7,15 @@ use config::node::{PeppygenLanguage, QoSProfile};
 use core_node::names;
 use core_node::{CoreNode, CoreNodeArguments};
 use core_node_api::encoding::{
-    NodeAddFeedback, NodeAddGoal, NodeAddGoalResponse, NodeAddResult, NodeBuildFeedback,
-    NodeBuildGoal, NodeBuildGoalResponse, NodeBuildResult, NodeRunFeedback, NodeRunGoal,
-    NodeRunGoalResponse, NodeRunResult, NodeSource,
+    ClockRequest, ClockResponse, ClockTick, NodeAddFeedback, NodeAddGoal, NodeAddGoalResponse,
+    NodeAddResult, NodeBuildFeedback, NodeBuildGoal, NodeBuildGoalResponse, NodeBuildResult,
+    NodeRunFeedback, NodeRunGoal, NodeRunGoalResponse, NodeRunResult, NodeSource, wall_now_ns,
 };
 use gix_url::Url as GitUrl;
 use node_stack::NodeStack;
-use peppylib::messaging::MessengerHandle;
+use peppylib::messaging::{MessengerHandle, TopicMessenger};
 use peppylib::runtime::{TaskHandle, spawn};
-use peppylib::{ActionMessenger, PeppyError};
+use peppylib::{ActionMessenger, PeppyError, ServiceMessenger};
 use pmi::{Messenger, MessengerAdapter, MessengerBackend, MockAdapter};
 use std::path::{Path, PathBuf};
 use std::process::{Command, Stdio};
@@ -85,6 +85,98 @@ pub async fn wait_until_service_reachable(
         }
         tokio::time::sleep(Duration::from_millis(10)).await;
     }
+}
+
+/// Drives the NTP-style 4-timestamp exchange against the started core node and
+/// asserts the wire contract: server echoes `t0` unchanged, and the causal
+/// chain `t0 ≤ t1 ≤ t2 ≤ t3` holds. Shared between the mock-messenger and
+/// real-zenoh round-trip tests.
+pub async fn assert_clock_round_trip(started: &StartedCoreNode) {
+    let t0 = wall_now_ns();
+    let request_payload = ClockRequest::new(t0)
+        .encode()
+        .expect("encode should succeed");
+
+    let response = ServiceMessenger::poll(
+        &started.caller_handle,
+        &started.core_node_name,
+        CALLER_INSTANCE_ID,
+        &started.core_node_name,
+        names::CLOCK,
+        Some(&started.core_node_name),
+        None,
+        request_payload,
+        Duration::from_secs(5),
+    )
+    .await
+    .expect("clock service poll should succeed");
+
+    let t3 = wall_now_ns();
+    let clock_response = ClockResponse::decode(&response.payload()).expect("decode should succeed");
+
+    assert_eq!(
+        clock_response.client_send_time, t0,
+        "server should echo client_send_time unchanged"
+    );
+    // Causal chain t0 ≤ t1 ≤ t2 ≤ t3 catches both unit mismatches (ns vs ms)
+    // and t1/t2 stamping-order regressions in one assert.
+    assert!(
+        t0 <= clock_response.server_recv_time
+            && clock_response.server_recv_time <= clock_response.server_send_time
+            && clock_response.server_send_time <= t3,
+        "expected t0 ({}) ≤ t1 ({}) ≤ t2 ({}) ≤ t3 ({})",
+        t0,
+        clock_response.server_recv_time,
+        clock_response.server_send_time,
+        t3,
+    );
+}
+
+/// Subscribes to the `clock` topic, collects three consecutive `ClockTick`s,
+/// and asserts they are strictly monotonic. Shared between the mock-messenger
+/// and real-zenoh publish tests.
+///
+/// Mock callers must pass `("*", "*")` for caller identity — the mock matcher
+/// is unidirectional and `emit_topic_message` hard-codes `*` into the
+/// caller-identity slots of the wire key. Real zenoh has bidirectional
+/// matching and accepts a real caller identity.
+pub async fn assert_clock_topic_emits_monotonic_ticks(
+    started: &StartedCoreNode,
+    caller_core_node: &str,
+    caller_instance_id: &str,
+    tick_timeout: Duration,
+) {
+    let mut subscription = TopicMessenger::subscribe(
+        &started.caller_handle,
+        caller_core_node,
+        caller_instance_id,
+        &started.core_node_name,
+        names::CLOCK,
+        Some(&started.core_node_name),
+        None,
+        QoSProfile::SensorData,
+    )
+    .await
+    .expect("clock topic subscription should succeed");
+
+    let mut times = Vec::with_capacity(3);
+    for _ in 0..3 {
+        let message = tokio::time::timeout(tick_timeout, subscription.on_next_message())
+            .await
+            .unwrap_or_else(|_| panic!("clock tick should arrive within {tick_timeout:?}"))
+            .expect("subscription should not close");
+
+        let tick = ClockTick::decode(message.payload().as_ref())
+            .expect("clock tick decode should succeed");
+        times.push(tick.time);
+    }
+
+    // Strict (not non-strict) so a publisher that re-emits the same payload
+    // doesn't silently pass.
+    assert!(
+        times.windows(2).all(|w| w[0] < w[1]),
+        "clock ticks should be strictly monotonic, got {times:?}",
+    );
 }
 
 fn init_test_data_dir() -> (TempDir, PeppyDirs) {
